requirements.txt
mariadb==1.1.14
python-dotenv==1.2.2
openpyxl==3.1.5
